make deal management screen
