+++ bilrun/.flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"image_picker","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/image_picker-0.6.7+22/","dependencies":[]},{"name":"multi_image_picker","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/multi_image_picker-4.8.00/","dependencies":[]},{"name":"url_launcher","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/url_launcher-6.0.3/","dependencies":[]}],"android":[{"name":"flutter_plugin_android_lifecycle","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/flutter_plugin_android_lifecycle-1.0.11/","dependencies":[]},{"name":"image_picker","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/image_picker-0.6.7+22/","dependencies":["flutter_plugin_android_lifecycle"]},{"name":"multi_image_picker","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/multi_image_picker-4.8.00/","dependencies":[]},{"name":"url_launcher","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/url_launcher-6.0.3/","dependencies":[]}],"macos":[{"name":"url_launcher_macos","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/url_launcher_macos-2.0.0/","dependencies":[]}],"linux":[{"name":"url_launcher_linux","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/url_launcher_linux-2.0.0/","dependencies":[]}],"windows":[{"name":"url_launcher_windows","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/url_launcher_windows-2.0.0/","dependencies":[]}],"web":[{"name":"url_launcher_web","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/url_launcher_web-2.0.0/","dependencies":[]}]},"dependencyGraph":[{"name":"flutter_plugin_android_lifecycle","dependencies":[]},{"name":"image_picker","dependencies":["flutter_plugin_android_lifecycle"]},{"name":"multi_image_picker","dependencies":[]},{"name":"url_launcher","dependencies":["url_launcher_linux","url_launcher_macos","url_launcher_windows","url_launcher_web"]},{"name":"url_launcher_linux","dependencies":[]},{"name":"url_launcher_macos","dependencies":[]},{"name":"url_launcher_web","dependencies":[]},{"name":"url_launcher_windows","dependencies":[]}],"date_created":"2021-04-27 13:43:59.100233","version":"2.0.5"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"image_picker","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/image_picker-0.6.7+22/","dependencies":[]},{"name":"multi_image_picker","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/multi_image_picker-4.8.00/","dependencies":[]},{"name":"url_launcher","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/url_launcher-6.0.3/","dependencies":[]}],"android":[{"name":"flutter_plugin_android_lifecycle","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/flutter_plugin_android_lifecycle-1.0.11/","dependencies":[]},{"name":"image_picker","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/image_picker-0.6.7+22/","dependencies":["flutter_plugin_android_lifecycle"]},{"name":"multi_image_picker","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/multi_image_picker-4.8.00/","dependencies":[]},{"name":"url_launcher","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/url_launcher-6.0.3/","dependencies":[]}],"macos":[{"name":"url_launcher_macos","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/url_launcher_macos-2.0.0/","dependencies":[]}],"linux":[{"name":"url_launcher_linux","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/url_launcher_linux-2.0.0/","dependencies":[]}],"windows":[{"name":"url_launcher_windows","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/url_launcher_windows-2.0.0/","dependencies":[]}],"web":[{"name":"url_launcher_web","path":"/Users/kimwonhee/.pub-cache/hosted/pub.dartlang.org/url_launcher_web-2.0.0/","dependencies":[]}]},"dependencyGraph":[{"name":"flutter_plugin_android_lifecycle","dependencies":[]},{"name":"image_picker","dependencies":["flutter_plugin_android_lifecycle"]},{"name":"multi_image_picker","dependencies":[]},{"name":"url_launcher","dependencies":["url_launcher_linux","url_launcher_macos","url_launcher_windows","url_launcher_web"]},{"name":"url_launcher_linux","dependencies":[]},{"name":"url_launcher_macos","dependencies":[]},{"name":"url_launcher_web","dependencies":[]},{"name":"url_launcher_windows","dependencies":[]}],"date_created":"2021-04-28 09:40:55.377603","version":"2.0.5"}
