+++ bilrun/lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:bilrun/screens/mypage/deal_list.dart';
+import 'package:bilrun/screens/mypage/deal_managemenet/deal_list.dart';
 import 'package:bilrun/screens/product_register/upload_file.dart';
 import 'package:bilrun/widgets/multi_image_picker.dart';
 import 'package:flutter/material.dart';
